reenable 2.13-only tests
