+++ language-support/scala/codegen-testing/src/test/scala/com/digitalasset/ledger/client/binding/PrimitiveSpec.scala
@@ -118,17 +118,6 @@ class PrimitiveSpec extends AnyWordSpec with Matchers with ScalaCheckDrivenPrope
       }
     }
 
-    /* 2.13 only
-    "converting with .to(GenMap)" should {
-      "preserve identity" in {
-        val gm = mii.to(GenMap)
-        isExactly(gm, ofType[GenMap[Int, Int]])
-        gm should ===(mii)
-        (gm eq mii) should ===(mii.to(Map) eq mii)
-      }
-    }
-     */
-
     "converting with from" should {
       "preserve identity" in {
         val gm = GenMap from mii
@@ -139,15 +128,6 @@ class PrimitiveSpec extends AnyWordSpec with Matchers with ScalaCheckDrivenPrope
     }
 
     ".map" should {
-
-      /* 2.13 only
-  "preserve type" in {
-      val gm: GenMap[Int, Int] = mii
-    val mapped = gm map identity
-    isExactly(mapped, ofType[GenMap[Int, Int]])
-  }
-       */
-
       import org.scalatest.matchers.dsl.ResultOfATypeInvocation
       import scala.collection.immutable.{HashMap, ListMap}
       Seq[(ResultOfATypeInvocation[_ <: Map[_, _]], GenMap[Int, Int])](
@@ -181,7 +161,6 @@ object PrimitiveSpec {
   private final class Proxy[A](val ignore: Unit) extends AnyVal
   // test conformance while disabling implicit conversion
   private def ofType[T]: Proxy[T] = new Proxy(())
-  // private def is[Ac, Ex](ac: Ac, ex: Proxy[Ex])(implicit conforms: Ac <:< Ex): Unit = ()
   // as a rule, the *singleton* type ac.type will not be ~ Ex; we are interested
   // in what expression `ac` infers to *absent context*.
   @silent("parameter value (ex|ac|ev) .* is never used")
